Implement LXPatternEngine scaffolding, prepping for pattern rack
