+++ src/main/java/heronarts/lx/midi/surface/APC40.java
@@ -41,6 +41,7 @@ import heronarts.lx.mixer.LXChannel;
 import heronarts.lx.mixer.LXAbstractChannel;
 import heronarts.lx.mixer.LXGroup;
 import heronarts.lx.mixer.LXMixerEngine;
+import heronarts.lx.mixer.LXPatternEngine;
 import heronarts.lx.parameter.BooleanParameter;
 import heronarts.lx.parameter.DiscreteParameter;
 import heronarts.lx.parameter.EnumParameter;
@@ -231,7 +232,7 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
     }
 
     private boolean isPatternEnabled(LXPattern pattern) {
-      switch (pattern.getChannel().compositeMode.getEnum()) {
+      switch (pattern.getChannel().patternEngine.compositeMode.getEnum()) {
       case BLEND:
         return pattern.enabled.isOn();
       default:
@@ -334,8 +335,8 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
     void onDeviceOnOff() {
       if (this.device instanceof LXPattern) {
         final LXPattern pattern = (LXPattern) this.device;
-        final LXChannel channel = pattern.getChannel();
-        if (channel.compositeMode.getEnum() == LXChannel.CompositeMode.BLEND) {
+        final LXPatternEngine engine = pattern.getPatternEngine();
+        if (engine.compositeMode.getEnum() == LXPatternEngine.CompositeMode.BLEND) {
           pattern.enabled.toggle();
         } else {
           pattern.getChannel().goPatternIndex(pattern.getIndex());
@@ -457,7 +458,7 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
       this.channel = channel;
       if (channel instanceof LXChannel c) {
         c.addListener(this);
-        c.compositeMode.addListener(this.onCompositeModeChanged);
+        c.patternEngine.compositeMode.addListener(this.onCompositeModeChanged);
       } else {
         channel.addListener(this);
       }
@@ -469,7 +470,7 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
       channel.stopClips.pending.addListener(this);
       channel.hasRunningClip.addListener(this);
       if (channel instanceof LXChannel c) {
-        c.focusedPattern.addListener(this);
+        c.patternEngine.focusedPattern.addListener(this);
         c.patterns.forEach(pattern -> this.patternListeners.put(pattern, new PatternListener(pattern)));
       }
       for (LXClip clip : this.channel.clips) {
@@ -482,7 +483,7 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
     public void dispose() {
       if (this.channel instanceof LXChannel c) {
         c.removeListener(this);
-        c.compositeMode.removeListener(this.onCompositeModeChanged);
+        c.patternEngine.compositeMode.removeListener(this.onCompositeModeChanged);
       } else {
         this.channel.removeListener(this);
       }
@@ -494,7 +495,7 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
       this.channel.stopClips.pending.removeListener(this);
       this.channel.hasRunningClip.removeListener(this);
       if (this.channel instanceof LXChannel c) {
-        c.focusedPattern.removeListener(this);
+        c.patternEngine.focusedPattern.removeListener(this);
       }
       this.patternListeners.values().forEach(patternListener -> patternListener.dispose());
       this.patternListeners.clear();
@@ -565,7 +566,7 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
         }
       }
       if (this.channel instanceof LXChannel c) {
-        if (p == c.focusedPattern) {
+        if (p == c.patternEngine.focusedPattern) {
           sendChannelPatterns(index, c);
         }
       }
@@ -827,7 +828,7 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
       final int endIndex = channel.patterns.size() - baseIndex;
       final int activeIndex = channel.getActivePatternIndex() - baseIndex;
       final int nextIndex = channel.getNextPatternIndex() - baseIndex;
-      final int focusedIndex = (channel.patterns.size() == 0) ? -1 : channel.focusedPattern.getValuei() - baseIndex;
+      final int focusedIndex = (channel.patterns.size() == 0) ? -1 : channel.patternEngine.focusedPattern.getValuei() - baseIndex;
 
       for (int y = 0; y < CLIP_LAUNCH_ROWS; ++y) {
         int pitch = CLIP_LAUNCH + y;
@@ -1099,8 +1100,8 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
       case BANK_SELECT_UP:
         if (this.shiftOn) {
           bus = lx.engine.mixer.getFocusedChannel();
-          if (bus instanceof LXChannel) {
-            ((LXChannel) bus).focusedPattern.decrement(1, false);
+          if (bus instanceof LXChannel channel) {
+            channel.patternEngine.focusedPattern.decrement(1, false);
           }
         } else {
           this.mixerSurface.decrementGridOffset();
@@ -1109,8 +1110,8 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
       case BANK_SELECT_DOWN:
         if (this.shiftOn) {
           bus = lx.engine.mixer.getFocusedChannel();
-          if (bus instanceof LXChannel) {
-            ((LXChannel) bus).focusedPattern.increment(1, false);
+          if (bus instanceof LXChannel channel) {
+            channel.patternEngine.focusedPattern.increment(1, false);
           }
         } else {
           this.mixerSurface.incrementGridOffset();
@@ -1158,9 +1159,9 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
             if (channel instanceof LXChannel c) {
               final int patternIndex = index + this.mixerSurface.getGridPatternOffset();
               if (patternIndex < c.patterns.size()) {
-                c.focusedPattern.setValue(patternIndex);
+                c.patternEngine.focusedPattern.setValue(patternIndex);
                 if (!this.shiftOn) {
-                  if (c.compositeMode.getEnum() == LXChannel.CompositeMode.BLEND) {
+                  if (c.patternEngine.compositeMode.getEnum() == LXPatternEngine.CompositeMode.BLEND) {
                     c.getPattern(patternIndex).enabled.toggle();
                   } else {
                     c.getPattern(patternIndex).launch.trigger();
@@ -1218,8 +1219,8 @@ public class APC40 extends LXMidiSurface implements LXMidiSurface.Bidirectional 
       return;
     case CHANNEL_FOCUS:
       if (this.shiftOn) {
-        if (channel instanceof LXChannel) {
-          ((LXChannel) channel).autoCycleEnabled.toggle();
+        if (channel instanceof LXChannel c) {
+          c.patternEngine.autoCycleEnabled.toggle();
         }
       } else {
         this.lx.engine.mixer.focusedChannel.setValue(channel.getIndex());
